Improve tests code structure slightly
